Separate the tests and benchmarks in datamodel-tests
