+++ pkg/datamodel-tests/src/main.bench.ts
@@ -5,7 +5,7 @@ import { newTestRunner } from '@runner/runner';
 import * as testDataLoader from '@loaders/couchdb-image-loader';
 
 import * as implementations from './implementations';
-import * as tests from './tests';
+import * as tests from './benchmarks';
 
 describe('Datamodel benchmarks', async () => {
 	const runner = await newTestRunner(testDataLoader);
